refactor: align visualization folder structure
