+++ libs/ui-react-visualization/src/lib/Components/LineChart/LineChart.tsx
@@ -1,4 +1,4 @@
-import type { Series } from '../../../utils';
+import type { Series } from '../../utils';
 
 type LineChartProps = {
   series: Series[];
